chore(client): adopt new protocol
- adjust create task command
- remove request/response from other commands (restore later)

related to #95

+++ client-java/src/main/java/org/camunda/tngp/client/cmd/CreateAsyncTaskCmd.java
@@ -2,7 +2,7 @@ package org.camunda.tngp.client.cmd;
 
 public interface CreateAsyncTaskCmd extends SetPayloadCmd<Long, CreateAsyncTaskCmd>
 {
-    CreateAsyncTaskCmd taskQueueId(int taskQueueId);
+    CreateAsyncTaskCmd taskQueueId(long taskQueueId);
 
     CreateAsyncTaskCmd taskType(String taskType);
 }
